requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==4.2.18
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-js-asset==3.0.1
django_csp==3.8
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2024.2
webencodings==0.5.1
django-axes==7.0.1
python-dotenv==1.0.1
whitenoise==6.8.2
django-check-seo==1.0.1
django-dbbackup==4.2.1
django-unfold==0.46.0
gunicorn==23.0.0
pillow==11.1.0
django-tailwind==3.8.0
django-cors-headers
dj-rest-auth
djangorestframework-simplejwt
coreschema
drf-yasg
django-filter
uritemplate
coreapi
pyyaml
drf-yasg[validation]
django-coreapi
openapi-codec
